Create Rate Initial Model
Now user can be an Employer and/or an Employee. Employee can provide service in a specific `workingRegion`. In the future we can create a Working Model and allow Employee to choose multiple regions.
Create a new policy `insertUserId()` that expect that the user is logged and add the userId from session to body.param. Used when you need to POST the logged userId.
Create the `view/becomeEmployee.ejs` inside private area. This form creates an Employee in a logged User.

Create Employee and Employer [Model and Controller]
model/User.js: Create Employer and Employee *unique* association; Automatcaly create an Employer in `afterCreate()`